Messaging Update (BUGGY)
Reconfigured how the connections are established on the client
end of things. Client was unable to receive messages due to
half-duplex. Threading has been enabled for the client in order to
provide a full duplex. Metadata of the server has beenchanged to
better keep track of connections and messages being sent. A method
has been created and main table has been altered to support new
messaging functionality as well as being able to print open connections.
Functionality for contact names has been started but not fully
developed. Closing of the server has been changed and needs to be
maintained in order to avoid errors.